Add ability to exclude SCSS files
Monkey patch the SCSSLint gem to expand the file path
to the same directory as the config file.
Since the config file is a Tempfile,
that directory can be acquired by creating another Tempfile

This only makes the top level `exclude:` option work
in the SCSSLint configuration file.

`exclude` at each individual SCSS linters level will not work
because we are passing SCSSLint string (file content),
and thus file path is not available when SCSSLint needs it.